+++ ignite/services/chain/generate.go
@@ -108,7 +108,7 @@ func (c *Chain) Generate(
 		return err
 	}
 
-	c.ev.Send("Building proto...", events.ProgressStarted())
+	c.ev.Send("Building proto...", events.ProgressUpdate())
 
 	options := []cosmosgen.Option{
 		cosmosgen.IncludeDirs(conf.Build.Proto.ThirdPartyPaths),
@@ -190,7 +190,7 @@ func (c *Chain) Generate(
 			c.ev.Send(
 				fmt.Sprintf("Typescript client path: %s", tsClientPath),
 				events.Icon(icons.Bullet),
-				events.ProgressFinished(),
+				events.ProgressFinish(),
 			)
 		}
 
@@ -198,7 +198,7 @@ func (c *Chain) Generate(
 			c.ev.Send(
 				fmt.Sprintf("Vuex stores path: %s", vuexPath),
 				events.Icon(icons.Bullet),
-				events.ProgressFinished(),
+				events.ProgressFinish(),
 			)
 		}
 
@@ -206,7 +206,7 @@ func (c *Chain) Generate(
 			c.ev.Send(
 				fmt.Sprintf("OpenAPI path: %s", openAPIPath),
 				events.Icon(icons.Bullet),
-				events.ProgressFinished(),
+				events.ProgressFinish(),
 			)
 		}
 	}